requirements.txt
GitPython==3.1.31
GitPython==3.1.31
openai==0.27.8
radon==6.0.1
Requests==2.31.0
streamlit==1.20.0
